third_party/blink/web_tests/accessibility/element-role-mapping-focusable-expected.txt

Visit Chromium Click the text! Button
Copyright 2015.
   I have a checkbox  Radio  Button
  I have a menuItem  Menu  ItemRadio
Birthday:  Birthday(date and time):  Birthday(datetime-local):  Birthday (month and year):  Select a week:  Select a file:  Min-Max:  Color: Select a time:   
Pacific
Atlantic
Indian
Southern
Arctic
 
Milk
Soda
Coffee
Water
  
0 100 + =
Blue  Red 
Text Input Controls
Text: 
Password: 


This test make sure that focusable elements are mapped implicitly to AX roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

normal_link_id
PASS elem.role is "AXRole: AXLink"
mousedown_link_id
PASS elem.role is "AXRole: AXLink"
button_id
PASS elem.role is "AXRole: AXButton"
details_id
PASS elem.role is "AXRole: AXDetails"
summary_id
PASS elem.role is "AXRole: AXDisclosureTriangle"
input_datalist_id
PASS elem.role is "AXRole: AXTextFieldWithComboBox"
input_button_id
PASS elem.role is "AXRole: AXButton"
input_checkbox_id
PASS elem.role is "AXRole: AXCheckBox"
input_radio_id
PASS elem.role is "AXRole: AXRadioButton"
menu_button_id
PASS elem.role is "AXRole: AXButton"
menu_checkbox_id
PASS elem.role is "AXRole: AXCheckBox"
menu_radio_id
PASS elem.role is "AXRole: AXRadioButton"
input_date_id
PASS elem.role is "AXRole: AXDate"
input_datetime_id
PASS elem.role is "AXRole: AXTextField"
input_datetime-local_id
PASS elem.role is "AXRole: AXDateTime"
input_month_id
PASS elem.role is "AXRole: AXDateTime"
input_week_id
PASS elem.role is "AXRole: AXDateTime"
input_file_id
PASS elem.role is "AXRole: AXButton"
input_number_id
PASS elem.role is "AXRole: AXSpinButton"
input_range_id
PASS elem.role is "AXRole: AXSlider"
input_color_id
PASS elem.role is "AXRole: AXColorWell"
input_time_id
PASS elem.role is "AXRole: AXInputTime"
input_reset_id
PASS elem.role is "AXRole: AXButton"
select_id
PASS elem.role is "AXRole: AXComboBoxSelect"
PASS axMenuListPopup.role is "AXRole: AXMenuListPopup"
select_option_id
PASS axMenuListOption.role is "AXRole: AXMenuListOption"
select_multiple_id
PASS elem.role is "AXRole: AXListBox"
select_multiple_option_id
PASS elem.role is "AXRole: AXListBoxOption"
textarea_id
PASS elem.role is "AXRole: AXTextField"
meter_id
PASS elem.role is "AXRole: AXMeter"
output_id
PASS elem.role is "AXRole: AXStatus"
input_radio_id
PASS elem.role is "AXRole: AXRadioButton"
legend_id
PASS elem.role is "AXRole: AXLegend"
input_text_id
PASS elem.role is "AXRole: AXTextField"
input_password_id
PASS elem.role is "AXRole: AXTextField"
PASS successfullyParsed is true

TEST COMPLETE
